Revert "Remove SendMediaReady legacy helper."
This reverts commit 91f8989f701c1c3bfe299ac8e653efc7d22a38e3.

+++ Telegram/SourceFiles/iv/iv_data.cpp
@@ -25,7 +25,7 @@ QByteArray GeoPointId(Geo point) {
 	const auto lon = int(point.lon * 1000000);
 	const auto combined = (std::uint64_t(std::uint32_t(lat)) << 32)
 		| std::uint64_t(std::uint32_t(lon));
-	return QByteArray::number(quint64(combined))
+	return QByteArray::number(combined)
 		+ ','
 		+ QByteArray::number(point.access);
 }
